Fix for Initial Startup of Driver
Kernel warning dispatched because, so it seems,
the device is sending a channel notify message
when the wiphy structure, on the Kernel side, isn't
quite ready to handle that event. This is hopefuly
a fix for that, inspired by this commit

https://github.com/RinCat/RTL88x2BU-Linux-Driver/tree/882dd86e929d2ffcc4027e8d08c605155053bb75/core

found in a later version of a similar driver by
Realtek.